using develop's new features and architecture while keeping master's
fixes and user-facing behaviour.

- `--is-reflink` exit code 5 means MAYBE_REFLINK in master, and
  INLINE_EXTENTS in develop. MAYBE_REFLINK stays 5 and INLINE_EXTENTS
  become 12. In the man page, take develop's phrasing for exit code 6
  as we got it wrong in 01f5aadd (#636). Sorry.
  Also, exit code 1 is RM_LINK_ERROR in develop but RM_LINK_NONE
  in master and vice versa. I kept master's to avoid breaking user
  scripts. Note that master returns EXIT_FAILURE(1) in some cases.

- kept `--write-unfinished` (as people using the master branch seem
  to like its speed), but as it can produce surprising results,
  its short option `-U` is changed to the the quasi-alternative
  `--hash-unmatched`. Updated tests to cover both. Added a note in man.

- kept `--without-compile-glib-schemas` over `--without-schemas-compile`.

- adapted several fixes from master to the new develop codebase

- updated gettext strings and po files

+++ lib/hasher.h
@@ -140,9 +140,9 @@ RmHasherTask *rm_hasher_task_new(RmHasher *hasher,
 gboolean rm_hasher_task_hash(RmHasherTask *task,
                              char *path,
                              guint64 start_offset,
-                             gsize bytes_to_read,
+                             guint64 bytes_to_read,
                              gboolean is_symlink,
-                             gsize *bytes_read_out);
+                             guint64 *bytes_read_out);
 
 /**
  * @brief Finalise a hashing task
